* develop: (5561 commits)
  timesheets2pdf.xml: first col enlarged for label structure element
  AuftragDO: statistics refactored (not yet invoiced, to be invoiced, invoiced etc.)
  AuftragsCache: zuFakturierenSum = beauftragt - fakturiert
  I18nKey.json: sort order of files and classes.
  Classical version: TimesheetEventsProvider: time zone handling.
  AuftragsCache works now with order.id=null for test cases.
  OrderExportTest fixed.
  OrderBook: statistics and to be invoiced values improved.
  Nothing (reformatted code before modificaiton)
  i18n key fixed.
  I18nKeysSourceAnalyzer improved.
  MyFullCalenderConfig revertet (timezone should be ignored).
  Changes of npm-updates-2 restored.
  Revert "jsx: eslint warnings."
  I18nKeyWIP: CreateI18nKeys refactored.
  jsx: eslint warnings.
  AuftragsCache for much better performance of order book's list view.
  AbstractListPage: long-running queries: change of sort order shows result list without search button. BeanHelper: lots of caches for faster sort in Wicket result lists. Eingangsrechnungslist: sort for konto fixed.
  CookieService: Update user's last login timestamp on using stay-logged-in-cookie.
  2FA: WebAuthn is not experimental anymore. Now, at least one instead of two different 2FA should be configured by all users.
  ...

# Conflicts:
#	plugins/org.projectforge.plugins.liquidityplanning/src/main/java/org/projectforge/plugins/liquidityplanning/LiquidityEntryListForm.java
#	plugins/org.projectforge.plugins.marketing/src/main/java/org/projectforge/plugins/marketing/AddressCampaignValueListForm.java
#	plugins/org.projectforge.plugins.skillmatrix/src/main/java/org/projectforge/plugins/skillmatrix/SkillSelectAutoCompleteFormComponent.java
#	plugins/org.projectforge.plugins.skillmatrix/src/main/java/org/projectforge/plugins/skillmatrix/SkillTreeExpansion.java
#	plugins/org.projectforge.plugins.skillmatrix/src/main/java/org/projectforge/plugins/skillmatrix/SkillTreeForm.java
#	plugins/org.projectforge.plugins.todo/src/main/java/org/projectforge/plugins/todo/ToDoEditForm.java
#	pom.xml
#	projectforge-wicket/src/main/java/de/micromata/less/LessWicketApplicationInstantiator.java
#	projectforge-wicket/src/main/java/net/ftlines/wicket/fullcalendar/FullCalendar.java
#	projectforge-wicket/src/main/java/net/ftlines/wicket/fullcalendar/callback/AbstractCallback.java
#	projectforge-wicket/src/main/java/net/ftlines/wicket/fullcalendar/callback/DateRangeSelectedCallback.java
#	projectforge-wicket/src/main/java/net/ftlines/wicket/fullcalendar/callback/EventClickedCallback.java
#	projectforge-wicket/src/main/java/net/ftlines/wicket/fullcalendar/callback/EventDroppedCallback.java
#	projectforge-wicket/src/main/java/net/ftlines/wicket/fullcalendar/callback/EventResizedCallback.java
#	projectforge-wicket/src/main/java/net/ftlines/wicket/fullcalendar/callback/ViewDisplayCallback.java
#	projectforge-wicket/src/main/java/org/projectforge/web/address/AddressCompareForm.java
#	projectforge-wicket/src/main/java/org/projectforge/web/address/AddressListForm.java
#	projectforge-wicket/src/main/java/org/projectforge/web/address/AddressListPage.java
#	projectforge-wicket/src/main/java/org/projectforge/web/admin/SystemUpdateForm.java
#	projectforge-wicket/src/main/java/org/projectforge/web/calendar/CalendarForm.java
#	projectforge-wicket/src/main/java/org/projectforge/web/core/importstorage/AbstractImportStoragePanel.java
#	projectforge-wicket/src/main/java/org/projectforge/web/fibu/AbstractRechnungEditForm.java
#	projectforge-wicket/src/main/java/org/projectforge/web/fibu/AbstractRechnungListForm.java
#	projectforge-wicket/src/main/java/org/projectforge/web/fibu/AuftragEditForm.java
#	projectforge-wicket/src/main/java/org/projectforge/web/fibu/RechnungListForm.java
#	projectforge-wicket/src/main/java/org/projectforge/web/gantt/GanttChartEditForm.java
#	projectforge-wicket/src/main/java/org/projectforge/web/gantt/GanttChartEditTreeTablePanel.java
#	projectforge-wicket/src/main/java/org/projectforge/web/meb/MebEditForm.java
#	projectforge-wicket/src/main/java/org/projectforge/web/task/TaskSelectAutoCompleteFormComponent.java
#	projectforge-wicket/src/main/java/org/projectforge/web/task/TaskTreeExpansion.java
#	projectforge-wicket/src/main/java/org/projectforge/web/teamcal/admin/TeamCalListForm.java
#	projectforge-wicket/src/main/java/org/projectforge/web/timesheet/TimesheetEditForm.java
#	projectforge-wicket/src/main/java/org/projectforge/web/tree/TreeIconsActionPanel.java
#	projectforge-wicket/src/main/java/org/projectforge/web/user/UserEditForm.java
#	projectforge-wicket/src/main/java/org/projectforge/web/user/UserPrefEditForm.java
#	projectforge-wicket/src/main/java/org/projectforge/web/wicket/AbstractListForm.java
#	projectforge-wicket/src/main/java/org/projectforge/web/wicket/InMemoryPageStore.java
#	projectforge-wicket/src/main/java/org/projectforge/web/wicket/LambdaModel.java
#	projectforge-wicket/src/main/java/org/projectforge/web/wicket/components/DatePanel.java
#	projectforge-wicket/src/main/java/org/projectforge/web/wicket/components/DatePickerUtils.java
#	projectforge-wicket/src/main/java/org/projectforge/web/wicket/components/DropFileContainer.java
#	projectforge-wicket/src/main/resources/org/projectforge/web/wicket/scripts/projectforge.js

+++ projectforge-model/src/main/java/org/projectforge/model/rest/CalendarEventObject.java
@@ -3,7 +3,7 @@
 // Project ProjectForge Community Edition
 //         www.projectforge.org
 //
-// Copyright (C) 2001-2014 Kai Reinhard (k.reinhard@micromata.de)
+// Copyright (C) 2001-2022 Micromata GmbH, Germany (www.micromata.com)
 //
 // ProjectForge is dual-licensed.
 //
@@ -23,16 +23,14 @@
 
 package org.projectforge.model.rest;
 
-import java.lang.reflect.Field;
-import java.util.Date;
-
-import javax.xml.bind.annotation.XmlRootElement;
-
-import org.apache.commons.lang.StringUtils;
-import org.apache.commons.lang.builder.ReflectionToStringBuilder;
-
 import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
 import com.fasterxml.jackson.databind.annotation.JsonSerialize;
+import org.apache.commons.lang3.StringUtils;
+import org.apache.commons.lang3.builder.ReflectionToStringBuilder;
+
+import javax.xml.bind.annotation.XmlRootElement;
+import java.lang.reflect.Field;
+import java.util.Date;
 
 /**
  * For documentation please refer the ProjectForge-API: TeamEventDO object.
@@ -66,7 +64,7 @@ public class CalendarEventObject extends AbstractBaseObject
 
   private String reminderType, reminderUnit;
 
-  private Integer reminderDuration;
+  private Long reminderDuration;
 
   private String recurrenceRule, recurrenceExDate;
 
@@ -184,12 +182,12 @@ public class CalendarEventObject extends AbstractBaseObject
    * @return the reminderDuration
    * @see #getReminder()
    */
-  public Integer getReminderDuration()
+  public Long getReminderDuration()
   {
     return reminderDuration;
   }
 
-  public CalendarEventObject setReminderDuration(final Integer reminderDuration)
+  public CalendarEventObject setReminderDuration(final Long reminderDuration)
   {
     this.reminderDuration = reminderDuration;
     return this;
